search: Adjust panel height according to num of rows
We cannot benefit here from the cc_list_box_setup_scrolling ()
list-box helper function because we are also embedding other
widgets in the scrolled window, such as the box which contains
the panel usage instructions and the gear button.

In doing so, we manually set the "cc-scrolling-scrolled-window"
and "cc-max-row-visible" properties in the GtkListBox so the
cc_list_box_adjust_scrolling () helper function can properly
calculate the total height of the visible list box rows.

MAX_ROWS_VISIBLE is arbitrarily set to 7 instead of 5 because
the other children of the scrolled window are not computed in
the cc_list_box_adjust_scrolling () function. The two lines
GtkLabel height summed to its vertical margins should equal
something around the size of two rows (2 + 5 = 7).

This commit should be reverted when the new Control Center
resizable shell gets introduced.

https://bugzilla.gnome.org/show_bug.cgi?id=775958